Restore compatibility with perl 5.8.8.
There are a few instances in the new code base that uses new features
introduced in 5.10. Unfortunately, some distros such as RHEL5 still has only
perl 5.8.8 in its repo, so those changes render Webwork unable to boot.

This commit restores compatibility with perl 5.8.8 to the point where it'll at
least boot. This is done by replacing the '//' operator with '||'.